Remove explicit cast

+++ src/jtagmkII.c
@@ -2437,9 +2437,9 @@ static int jtagmkII_get_sck_period(const PROGRAMMER *pgm, double *v) {
   else if (buf[0] == 1)
     clk = 2.8e6;
   else if (buf[0] <= 5)
-    clk = 5.35e6 / (double)buf[0];
+    clk = 5.35e6 / buf[0];
   else
-    clk = 5.35e6 / (double)buf[0];
+    clk = 5.35e6 / buf[0];
 
   *v = 1 / clk;
   return 0;
